+++ tests/resource_error.py
@@ -1,4 +1,5 @@
 from rest_toolkit import resource
+from pyramid.httpexceptions import HTTPFound
 from pyramid.httpexceptions import HTTPNotFound
 from pyramid.httpexceptions import HTTPPaymentRequired
 
@@ -19,3 +20,9 @@ class HTTPErrorResource(object):
 class HTTPNotFoundResource(object):
     def __init__(self, request):
         raise HTTPNotFound()
+
+
+@resource('/http-found')
+class HTTPFoundResource(object):
+    def __init__(self, request):
+        raise HTTPFound('http://www.wiggy.net')